requirements.txt
playwright~=1.31.0
pytest-playwright
pytest~=7.2.0
pytest-html
pytest-rerunfailures
seldom~=3.1.3
allure-pytest==2.12.0
allure-python-commons==2.12.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
redis==4.3.5

anyio
pytest-asyncio
pytest-tornasync
pytest-trio
pytest-twisted

py~=1.11.0
python-slugify~=8.0.1